fix(server): re-tag output mismatches as internal
An RPC procedure with .output() let the raw ValidationError from a
handler-result mismatch escape to the client: code VALIDATION_ERROR is a
plain 400, so toErrorBody echoed the message verbatim — including the
concrete offending server-side value — and the server contract bug was
misreported as a client error, invisible to 5xx alerting.

Extract the REST path's applyOutput helper (which already re-tags the
mismatch as INTERNAL_SERVER_ERROR, redacted at the wire and logged
server-side) into src/apply-output.ts and route makeHandler through it.
Args validation failures are unchanged and stay 400s.

The existing builder test asserting ValidationError from an output
mismatch encoded the bug; it now asserts a LunoraError with code
INTERNAL_SERVER_ERROR.
